SOLR-17884: SolrJ Utils: Move org.apache.http usage to HttpClientUtil
