+++ solr/core/src/test/org/apache/solr/filestore/TestDistribFileStore.java
@@ -38,6 +38,7 @@ import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.SolrRequest;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.impl.BaseHttpSolrClient.RemoteExecutionException;
+import org.apache.solr.client.solrj.impl.HttpClientUtil;
 import org.apache.solr.client.solrj.impl.HttpSolrClient;
 import org.apache.solr.client.solrj.request.FileStoreApi;
 import org.apache.solr.client.solrj.request.V2Request;
@@ -182,7 +183,7 @@ public class TestDistribFileStore extends SolrCloudTestCase {
           j.getBaseURLV2() + "/cluster/filestore/files" + "/package/mypkg/v1.0/runtimelibs.jar";
       HttpDelete del = new HttpDelete(path);
       try (HttpSolrClient cl = (HttpSolrClient) j.newClient()) {
-        Utils.executeHttpMethod(cl.getHttpClient(), path, Utils.JSONCONSUMER, del);
+        HttpClientUtil.executeHttpMethod(cl.getHttpClient(), path, Utils.JSONCONSUMER, del);
       }
       expected = Collections.singletonMap(":files:/package/mypkg/v1.0/runtimelibs.jar", null);
       checkAllNodesForFile(cluster, "/package/mypkg/v1.0/runtimelibs.jar", expected, false);
@@ -205,7 +206,7 @@ public class TestDistribFileStore extends SolrCloudTestCase {
       if (verifyContent) {
         try (HttpSolrClient solrClient = (HttpSolrClient) jettySolrRunner.newClient()) {
           ByteBuffer buf =
-              Utils.executeGET(
+              HttpClientUtil.executeGET(
                   solrClient.getHttpClient(),
                   baseUrl + "/cluster/filestore/files" + path,
                   Utils.newBytesConsumer(Integer.MAX_VALUE));
@@ -230,7 +231,7 @@ public class TestDistribFileStore extends SolrCloudTestCase {
     public NavigableObject call() throws Exception {
       try (HttpSolrClient solrClient = (HttpSolrClient) jetty.newClient()) {
         return (NavigableObject)
-            Utils.executeGET(solrClient.getHttpClient(), this.url, JAVABINCONSUMER);
+            HttpClientUtil.executeGET(solrClient.getHttpClient(), this.url, JAVABINCONSUMER);
       }
     }
