+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/HttpSolrClient.java
@@ -721,7 +721,7 @@ public class HttpSolrClient extends BaseHttpSolrClient {
           "IOException occurred when talking to server at: " + getBaseURL(), e);
     } finally {
       if (shouldClose) {
-        Utils.consumeFully(entity);
+        HttpClientUtil.consumeFully(entity);
       }
     }
   }
